requirements.txt
streamlit
google.generativeai
python-dotenv
pymongo
scikit-learn
Flask-pymongo